Consolidate Slack app into response
