+++ response/admin.py
@@ -2,3 +2,4 @@ from django.contrib import admin
 
 from .core import admin
 from .pagerduty import admin
+from .slack import admin
