+++ response/models.py
@@ -2,3 +2,4 @@ from django.db import models
 
 from .core.models import *
 from .pagerduty.models import *
+from .slack.models import *
