+++ response/urls.py
@@ -1,2 +1,3 @@
 
 from .core.urls import *
+from .slack.urls import *
